Safari-611.1.10/LayoutTests/editing/pasteboard/paste-cocoa-writer-markup-with-webkit-standard-font-family-expected.txt

This tests pasting a markup with -webkit-standard font family name. WebKit should not strip it away upon paste.
To manually test, click on "start" first then trigger paste from menu or callout bar or press cmd+v

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.clipboardData.getData('text/html') is markup
PASS editor.innerHTML.includes('-webkit-standard') is true
PASS successfullyParsed is true

TEST COMPLETE
